refactor(readiness): dedupe probe-FS helpers and slim probes
