+++ src/integration/ai/readiness/claude/probe.ts
@@ -1,13 +1,16 @@
-import { promises as fs, type Stats } from 'node:fs';
-import { basename, join } from 'node:path';
+import { join } from 'node:path';
 import { Result } from '@src/domain/result.ts';
-import { type AbsolutePath } from '@src/domain/value/absolute-path.ts';
-import { Slug } from '@src/domain/value/slug.ts';
-import { toKebabCase } from '@src/domain/value/kebab-case.ts';
+import type { AbsolutePath } from '@src/domain/value/absolute-path.ts';
 import type { IsoTimestamp } from '@src/domain/value/iso-timestamp.ts';
 import { ProbeError } from '@src/domain/value/error/probe-error.ts';
 import type { Repository } from '@src/domain/entity/repository.ts';
-import type { ArtifactRef, HookRef, NamedArtifactRef } from '@src/integration/ai/readiness/_engine/artifact-ref.ts';
+import type { ArtifactRef, HookRef } from '@src/integration/ai/readiness/_engine/artifact-ref.ts';
+import {
+  probeFile,
+  probeNamedDirCollection,
+  probeNamedFileCollection,
+  readFileSafely,
+} from '@src/integration/ai/readiness/_engine/probe-fs.ts';
 import type { ClaudeArtifacts } from '@src/integration/ai/readiness/claude/artifacts.ts';
 import { absentState, type ReadinessState, presentState } from '@src/integration/ai/readiness/_engine/state.ts';
 import { hasAnyClaudeArtifact } from '@src/integration/ai/readiness/_engine/predicates.ts';
@@ -76,91 +79,6 @@ export const claudeProbe: ReadinessProbe<ClaudeArtifacts> = {
   },
 };
 
-const probeFile = async (path: string): Promise<Result<ArtifactRef | undefined, ProbeError>> => {
-  try {
-    const stat = await fs.stat(path);
-    if (!stat.isFile()) return Result.ok(undefined);
-    return Result.ok({ path: path as AbsolutePath });
-  } catch (cause) {
-    if (isNodeErrnoCode(cause, 'ENOENT')) return Result.ok(undefined);
-    if (isNodeErrnoCode(cause, 'EACCES')) {
-      return Result.error(
-        new ProbeError({ subCode: 'fs-permission', message: `permission denied reading ${path}`, path, cause })
-      );
-    }
-    return Result.error(new ProbeError({ subCode: 'fs-read', message: `failed to stat ${path}`, path, cause }));
-  }
-};
-
-const probeNamedFileCollection = async (dir: string): Promise<Result<NamedArtifactRef[], ProbeError>> => {
-  const entries = await listDir(dir);
-  if (!entries.ok) return Result.error(entries.error);
-  const refs: NamedArtifactRef[] = [];
-  for (const entry of entries.value) {
-    if (!entry.endsWith('.md')) continue;
-    const full = join(dir, entry);
-    const stat = await statSafely(full);
-    if (!stat.ok) return Result.error(stat.error);
-    if (stat.value === undefined || !stat.value.isFile()) continue;
-    const baseName = entry.slice(0, -'.md'.length);
-    const slug = Slug.parse(toKebabCase(baseName));
-    if (!slug.ok) continue;
-    refs.push({ name: slug.value, path: full as AbsolutePath });
-  }
-  return Result.ok(refs);
-};
-
-const probeNamedDirCollection = async (
-  dir: string,
-  childMarker: string
-): Promise<Result<NamedArtifactRef[], ProbeError>> => {
-  const entries = await listDir(dir);
-  if (!entries.ok) return Result.error(entries.error);
-  const refs: NamedArtifactRef[] = [];
-  for (const entry of entries.value) {
-    const childDir = join(dir, entry);
-    const stat = await statSafely(childDir);
-    if (!stat.ok) return Result.error(stat.error);
-    if (stat.value === undefined || !stat.value.isDirectory()) continue;
-    const markerPath = join(childDir, childMarker);
-    const markerStat = await statSafely(markerPath);
-    if (!markerStat.ok) return Result.error(markerStat.error);
-    if (markerStat.value === undefined || !markerStat.value.isFile()) continue;
-    const slug = Slug.parse(toKebabCase(basename(entry)));
-    if (!slug.ok) continue;
-    refs.push({ name: slug.value, path: markerPath as AbsolutePath });
-  }
-  return Result.ok(refs);
-};
-
-const listDir = async (dir: string): Promise<Result<string[], ProbeError>> => {
-  try {
-    return Result.ok(await fs.readdir(dir));
-  } catch (cause) {
-    if (isNodeErrnoCode(cause, 'ENOENT') || isNodeErrnoCode(cause, 'ENOTDIR')) return Result.ok([]);
-    if (isNodeErrnoCode(cause, 'EACCES')) {
-      return Result.error(
-        new ProbeError({ subCode: 'fs-permission', message: `permission denied listing ${dir}`, path: dir, cause })
-      );
-    }
-    return Result.error(new ProbeError({ subCode: 'fs-read', message: `failed to read ${dir}`, path: dir, cause }));
-  }
-};
-
-const statSafely = async (path: string): Promise<Result<Stats | undefined, ProbeError>> => {
-  try {
-    return Result.ok(await fs.stat(path));
-  } catch (cause) {
-    if (isNodeErrnoCode(cause, 'ENOENT')) return Result.ok(undefined);
-    if (isNodeErrnoCode(cause, 'EACCES')) {
-      return Result.error(
-        new ProbeError({ subCode: 'fs-permission', message: `permission denied stat ${path}`, path, cause })
-      );
-    }
-    return Result.error(new ProbeError({ subCode: 'fs-read', message: `failed to stat ${path}`, path, cause }));
-  }
-};
-
 const readHooks = async (
   settingsRefs: ReadonlyArray<ArtifactRef | undefined>
 ): Promise<Result<HookRef[], ProbeError>> => {
@@ -181,20 +99,6 @@ const readHooks = async (
     extractHooks(parsed, hooks);
   }
   return Result.ok(hooks);
-};
-
-const readFileSafely = async (path: AbsolutePath): Promise<Result<string | undefined, ProbeError>> => {
-  try {
-    return Result.ok(await fs.readFile(path, 'utf8'));
-  } catch (cause) {
-    if (isNodeErrnoCode(cause, 'ENOENT')) return Result.ok(undefined);
-    if (isNodeErrnoCode(cause, 'EACCES')) {
-      return Result.error(
-        new ProbeError({ subCode: 'fs-permission', message: `permission denied reading ${path}`, path, cause })
-      );
-    }
-    return Result.error(new ProbeError({ subCode: 'fs-read', message: `failed to read ${path}`, path, cause }));
-  }
 };
 
 /**
@@ -222,6 +126,3 @@ const extractHooks = (settings: unknown, sink: HookRef[]): void => {
     }
   }
 };
-
-const isNodeErrnoCode = (cause: unknown, code: string): boolean =>
-  typeof cause === 'object' && cause !== null && (cause as { code?: unknown }).code === code;
